refactor: Access is now always with type safe RegisterId

+++ src/libmicroemu/internal/executor/instr/ternary_load_instr_with_shift.h
@@ -53,7 +53,7 @@ public:
     if (is_wback == true) {
       Reg::WriteRegister(ictx.pstates, arg_n.Get(), offset_address);
     }
-    if (arg_t.Get() == static_cast<u8>(RegisterId::kPc)) {
+    if (arg_t.Get() == RegisterId::kPc) {
       // When the given address was unaligend the behaviour is
       // unpredtictable
       if ((address & 0x3u) == 0u) {
